+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.2.0-alpha.5 October 24, 2022
+- Fix document highlights for image links.
+- Make document highlights work for external links.
+
 ## 0.2.0-alpha.4 October 24, 2022
 - Fix document highlights for links to headers in other files.